Fix bug with lazy EODs resolving twice (#83)
If the client query includes a node, and a resolver RSS also includes the same node in the same OER (same path from the root query), we attempt to call `resolveData` on the lazyeod twice, which ends up calling the node resolver twice. This also results in OER.resolve() throwing since it's already been resolved. This fixes this bug, and also migrates NodeResolverTest from the old to the new engine feature test framework so I can add a test case there.
Github-Change-Id: 951240

GitOrigin-RevId: ac6e62c630c899e7468bfca9f8069252d9bbb12f

Co-authored-by: Aileen Chen <gummybug@users.noreply.github.com>