refactor: Replace ExecutionContext impls with ne... (#109)
PR #942164 (1/n) migrated to the new implementation of node execution contexts.  This current PR continues that work by migrating field execution contexts (both regular and mutation) as well as variable-provider contexts.  This was significantly more work than the previous one, and so this is a large PR (and it builds on quite a few before it).  The result of all of this is not only simpler factories but also testing infrastructure that is less coupled and should be easier to maintain.

There's one more PR coming in this series: we will remove the old `viaduct.runtime.context` plus a bunch of helper abstractions for that which are no longer used, and will rename `context2` to `context`.
Github-Change-Id: 953088
GitOrigin-RevId: 378eac12765dec77fe12a7128106109316691642

Co-authored-by: Raymie Stata <rstata@users.noreply.github.com>