refactor: Further consolidation of GRT factories (#100)
In transitioning from `context` to `context2` we continue to align the two _prior_ to the actual transition.  What we're discovering is that factories for GRT instances are a bit fractured.  And further, to support testing -- and especially testing with _out_ codegen (ie, "feature tests"), there are special-cases in the factories that, combined with the facturedness, impedes maintenance.

To address this issue, this PR further consolidates GRT factories into GRTConstructorExtensions.kt, which we introduced a while ago to help align `context` and `context2`.  An important element of this PR is a more systematic approach to "fake" GRTs.  These are GRTs that are _not_ generated by codegen, and therefore do not conform to the naming and other invariants of "official" GRTs, but are very helpful for testing without codegen.  These used to live in a file called `stubs.kt` and required mocking and other mechanisms to properly get injected into tests.  By giving these fakes first-class status in our GRT factories, we decouple the implementation of contexts from these helpful abstractions.
Github-Change-Id: 952709
GitOrigin-RevId: d0152af70a00c4adcdc0de031c58c8bc73576583

Co-authored-by: Raymie Stata <rstata@users.noreply.github.com>