refactor: Consolidate ViaductSchema and ViaductE... (#128)
The `viaduct.graphql.schema` package originally defined two schema APIs: `ViaductSchema` and `ViaductExtendedSchema`.  But as a practical matter, only `ViaductExtendedSchema` is the only one that was used, and having the two APIs has become a maintenance burden.  This PR eliminates the old `ViaductSchema` in favor of `ViaductExtendedSchema` - but then renames `ViaductExtendedSchema` to `ViaductSchema`.
Github-Change-Id: 955792

GitOrigin-RevId: e95061313bb72161c2fe17d8943d7917b5d47356

Co-authored-by: Raymie Stata <rstata@users.noreply.github.com>