throw `RequestScopeCancellationException` when we... (#197)
In #960503, we implemented cancellation of lingering jobs after the execution result was built and the request was "complete" from the perspective of the client. This change both a) prevents spurious timeouts and b) ensures that we don't perform work that isn't needed. However, this was causing `JobCancellationExceptions` to be thrown by fields that were cancelled, which leaked into our field error metrics.

This PR cancels our supervisor scope with a `RequestScopeCancellationException`, which we can then mark as "non-fatal" in order to not screw up our error reporting. Any other cancellation exceptions that happen (which are not expected) would still be marked as fatal errors.
Github-Change-Id: 961721
GitOrigin-RevId: 3d7687f2d4d7bc5db1521f0bcb99536611bb6793

Co-authored-by: Adam Miskiewicz <skevy@users.noreply.github.com>