EngineObjectData.Sync getSelections (#203)
To support object mapping, it would be useful if there were a synchronous method that returned the keys in an `EngineObjectData.Sync`.

`EngineObjectData.Sync` already supports synchronous versions of the suspending methods on EngineObjectData, *except* for a synchronous version of `fetchSelections`. This PR adds a synchronous `getSelections` method, which is in keeping with the spirit of this interface.

As a small (but significant!) change, this PR also changes `ResolvedEngineObjectData` into a data class. This helps with constructing these objects in places where we have a complete map in hand, and allows for reasonable value-based equality checks.
Github-Change-Id: 961473
GitOrigin-RevId: 20bcaadf3d54f3a57c929f798cca80639ce4b600

Co-authored-by: James Bellenger <jbellenger@users.noreply.github.com>