Refactor ResolverTestBase (#257)
## Summary
`ResolverTestBase` was implemented using the actual implementations of the various `ExecutionContext` interfaces.  Since the construction of the actual implementations is entangled with the bootstrapping process, this made it difficult to initialize this test fixture, difficult to inject test data into the resolver tests that this class is meant to support, and difficult to change the implementation details of `ExecutionContext.`. This PR switches `ResolverTestBase` to use test doubles for `ExecutionContext`, decoupling it from the actual implementations of `ExecutionContext`.

GitOrigin-RevId: 1c0656a6793819a85c835c2bcc8b783331486de9