fix: Ensures that executed checker RSSes are always part of the query plan (AIRBNB)
Consider the query `{ a }` against the following example schema:
```
type Query {
a: String # Has a checker with 'b' as its RSS
b: String # Has a checker with 'c' as its RSS
c: String
}
```
Since we bypass access checks for checker RSSes, we don't read the checker slot for `b` when executing this query, only its raw slot. We previously used this insight to avoid constructing the child QueryPlan for `b`'s checker as both an optimization and to avoid cycles. However, even though the checker slot is never read, it is written because whenever we resolve a field we always execute its raw and checker slots atomically. This results in timeouts because the checker for `b` never completes, waiting for `c` which is never written.

This PR fixes this with 2 changes:
1. Removes inCheckerContext, and always creates child plans for checkers for visited fields. Avoids cycles by instead tracking seen RSSes in each branch of the QueryPlan tree.
2. The above change dramatically increases the size of field type child plans we create for fields with interface types that can have many implementers like `node(id: ID!): Node`. However, we only ever execute one of these at runtime depending on what concrete type the field resolved to, which is typically the same across all executions of a given operation. So a second major change in this PR constructs the field type child plans lazily to avoid a combinatorial explosion. I also tested changing field child plans to be built lazily, but it didn't improve anything so I reverted that to reduce complexity.

Github-Change-Id: 975932
GitOrigin-RevId: a83674f0b003e8923bf3d48a4fb480f973a38c4c